refactor: constant index -> Group으로 변경
